+++ look/CHANGELOG.md
@@ -1,7 +1,13 @@
-## 0.0.1
+# look package changelog
 
-* Initial release
+## 1.0.1
+
+* Fix package description
 
 ## 1.0.0
 
 * First publication
+
+## 0.0.1
+
+* Initial release