+++ look_generator/CHANGELOG.md
@@ -1,8 +1,9 @@
-## 0.0.1
-
-* Initial release
+# look_generator package changelog
 
 ## 1.0.0
 
 * First publication
 
+## 0.0.1
+
+* Initial release